(#269) Move parse_assets_conf to something_parsers
And reuse it everywhere it's needed
