+++ src/something_assets.cpp
@@ -270,25 +270,6 @@ void Assets::clean()
     animats_count = 0;
 }
 
-// TODO: Function parse_assets_conf is copypasted from assets_typer.cpp
-template <typename F>
-void parse_assets_conf(String_View input, F f)
-{
-    for (int line_number = 1; input.count > 0; ++line_number) {
-        String_View line = input.chop_by_delim('\n').trim();
-
-        if (line.count == 0) continue; // Skip empty lines
-        if (*line.data == '#') continue; // Skip single line comments
-
-        String_View asset_type = line.chop_by_delim('[').trim();
-        String_View asset_id = line.chop_by_delim(']').trim();
-        line.chop_by_delim('=');
-        String_View asset_path = line.chop_by_delim('#').trim();
-
-        f(line_number, asset_type, asset_id, asset_path);
-    }
-}
-
 void Assets::load_conf(SDL_Renderer *renderer, const char *filepath)
 {
     clean();
